test_mission.pyのテストをtest_parameters2.pyに移動
